+++ lib/ui/assistant/AssistantConversationContentWidget.dart
@@ -50,8 +50,9 @@ import 'package:rokwire_plugin/utils/utils.dart';
 class AssistantConversationContentWidget extends StatefulWidget {
   final Stream shouldClearAllMessages;
   final AssistantProvider? provider;
+  final String? initialQuestion;
 
-  AssistantConversationContentWidget({required this.shouldClearAllMessages, this.provider});
+  AssistantConversationContentWidget({required this.shouldClearAllMessages, this.provider, this.initialQuestion});
 
   @override
   State<AssistantConversationContentWidget> createState() => _AssistantConversationContentWidgetState();
@@ -107,7 +108,9 @@ class _AssistantConversationContentWidgetState extends State<AssistantConversati
     });
 
     _loadLocationStatus();
-    _onPullToRefresh();
+    _onPullToRefresh().then((_) {
+      _askInitialQuestionIfAllowed();
+    });
 
     _userContext = _getUserContext();
 
@@ -1127,6 +1130,12 @@ class _AssistantConversationContentWidgetState extends State<AssistantConversati
 
   void _scrollListener() {
     _scrollPosition = _scrollController.position.pixels;
+  }
+
+  void _askInitialQuestionIfAllowed() {
+    if ((widget.initialQuestion != null) && (widget.initialQuestion?.isNotEmpty == true) && (_provider != null) && ((_availableQueryLimit ?? 0) > 0)) {
+      _submitMessage(message: widget.initialQuestion ?? '', provider: _provider!);
+    }
   }
 
   void _loadLocationStatus() {