+++ CHANGELOG.md
@@ -5,6 +5,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Added VideoTutorialThumbButton to Onboarding2GetStartedPanel [#4863](https://github.com/rokwire/illinois-app/issues/4863).
+### Fixed
+- Do not include "rejected" groups in user groups' count [#5207](https://github.com/rokwire/illinois-app/issues/5207).
+- Fixed address field in Digital Business Card generation [#5203](https://github.com/rokwire/illinois-app/issues/5203).
+### Changed
+- Share My Info moved to Profile Info page [#5201](https://github.com/rokwire/illinois-app/issues/5201).
+- Updated privacy icons in profile edit page [#5200](https://github.com/rokwire/illinois-app/issues/5200).
+- Updated Food Filters icon [#5206](https://github.com/rokwire/illinois-app/issues/5206).
+### Added
+- Assistant Events Prompt [#5220](https://github.com/rokwire/illinois-app/issues/5220).
 
 ## [7.2.0] - 2025-06-05
 ### Added